// example/src/component/utils/types/colors.ts
// Copyright 2020 Prescryptive Health, Inc.

export enum PurpleScale {
  light = '#EEE5FF',
  medium = '#7754B8',
  regular = '#6240A3',
  dark = '#4A2B79',
}

export enum PurpleGradient {
  dark = 'rgba(98,64,163,0.7)',
  light = 'rgba(98,64,163,0.2)',
}

export enum BlueScale {
  regular = '#2CBDE5',
  dark = '#00A3D9',
  light = '#E2F2F9',
}

export enum GreyScale {
  white = '#ffffff',
  light = '#F5F5FA',
  medium = '#CCCED9',
  dark = '#6C6F80',
  black = '#1B1F32',
  disabledIcon = '#CCCED9',
}

export enum RedScale {
  dark = '#99262A',
  light = '#E5454A',
  error = '#FF6660',
}

export enum YellowScale {
  dark = '#E6AC00',
}

export enum Transparency {
  dark = '#00000050',
}

export enum GreenScale {
  enabledIcon = '#54BF30',
  success = '#CAFFB9',
}